src/test/java/de/jensd/addon/Base64DecoderTest.java
package de.jensd.addon;

import java.util.List;
import java.util.Map;
import java.util.stream.Collectors;
import org.junit.Test;
import static org.junit.Assert.*;
import de.jensd.addon.decoder.PayloadDecoder;

/**
 *
 * @author Jens Deters
 */
public class Base64DecoderTest {

    @Test
    public void testBase64Decoder() {
        final String CONTENT = "Hello World";
        final byte[] PAYLOAD = CONTENT.getBytes();
        AddOnRegistryServiceLoader registry = new AddOnRegistryServiceLoader();
        List<PayloadDecoder> decoders = registry.getAddOns(PayloadDecoder.class);
        Map<String, PayloadDecoder> decodersMap = decoders.stream().collect(
                Collectors.toMap(c -> c.getId(), c -> c));
        PayloadDecoder decoder = decodersMap.get("base64_decoder");
        assertNotNull("Base64Decoder must not be null", decoder);
        String decoded = decoder.decode(PAYLOAD);
        assertEquals(decoded, "SGVsbG8gV29ybGQ=");
    }

}
